rag-pipeline/requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic
#pydantic-settings
python-dotenv==1.0.1
pydantic_settings

# PDF Parsing
pymupdf
pdfplumber==0.11.1

# NLP / Chunking
nltk==3.8.1
sentence-transformers==3.0.1
tiktoken

# IR Baselines
rank-bm25==0.2.2
scikit-learn
numpy==1.26.4

# Vector Store
chromadb

# OpenAI
openai

# Evaluation
scipy
# Utilities
tqdm
rich
httpx
